Restructure messages
