+++ src/client/app/messages/admin/admin-messages.module.ts
@@ -4,11 +4,12 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 
-import { UtilModule } from 'app/shared/util.module';
+import { UtilModule } from 'app/shared';
+
+import { MessagesModule } from '../messages.module';
 import { CreateMessageComponent } from './create-message.component';
 import { ListMessagesComponent } from './list-messages.component';
 import { AdminMessagesRoutingModule } from './admin-messages-routing.module';
-import { MessagesModule } from '../messages.module';
 import { UpdateMessageComponent } from './edit-message.component';
 
 @NgModule({
